Added bloom filter support.
* Filter block is one of meta blocks.
* Filter arrangements:
    1. BLOCK[0 ~ N-1]
    2. BLOCK_OFFSET[0 ~ N-1]
    3. OFFSET_OFFSET (4 bytes)
    4. Filter Base (1 byte)

In particular, we add a new FilterPolicy class.  An instance
of this class can be supplied in Options when opening a
database.  If supplied, the instance is used to generate
summaries of keys (e.g., a bloom filter) which are placed in
sstables.  These summaries are consulted by DB::Get() so we
can avoid reading sstable blocks that are guaranteed to not
contain the key we are looking for.

This change provides one implementation of FilterPolicy
based on bloom filters.

Other changes:
- Updated version number to 1.4.
- Some build tweaks.
- C binding for CompactRange.
- A few more benchmarks: deleteseq, deleterandom, readmissing, seekrandom.
- Minor .gitignore update.